refactor(user-profile): split edit profile to multiple components
